requirements.txt
python-telegram-bot==20.7
gspread==6.0.2
google-auth==2.27.0
python-dotenv==1.0.1
